% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxaca.R
\name{oaxaca}
\alias{oaxaca}
\title{Kitagawa–Blinder–Oaxaca decomposition
Author: Mediacom
This function allows you to run a Kitagawa–Blinder–Oaxaca decomposition.
oaxaca()}
\usage{
oaxaca(formula, data, group.weights = NULL, R = 100, reg.fun = lm, ...)
}
\description{
Kitagawa–Blinder–Oaxaca decomposition
Author: Mediacom
This function allows you to run a Kitagawa–Blinder–Oaxaca decomposition.
oaxaca()
}
