% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxaca.R
\name{plot.oaxaca}
\alias{plot.oaxaca}
\title{Kitagawa–Blinder–Oaxaca decomposition
Author: Mediacom
This function allows you to plot your results.
plot.oaxaca()}
\usage{
\method{plot}{oaxaca}(
  x,
  decomposition = "threefold",
  type = "variables",
  group.weight = NULL,
  unexplained.split = FALSE,
  variables = NULL,
  components = NULL,
  component.left = FALSE,
  component.labels = NULL,
  variable.labels = NULL,
  ci = TRUE,
  ci.level = 0.95,
  title = "",
  xlab = "",
  ylab = "",
  bar.color = NULL,
  ...
)
}
\description{
Kitagawa–Blinder–Oaxaca decomposition
Author: Mediacom
This function allows you to plot your results.
plot.oaxaca()
}
