% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxaca.R
\name{summary.oaxaca}
\alias{summary.oaxaca}
\title{Kitagawa–Blinder–Oaxaca decomposition
Author: Mediacom
This function allows you to run a summary of Kitagawa–Blinder–Oaxaca decomposition.
summary.oaxaca()}
\usage{
\method{summary}{oaxaca}(object, ...)
}
\description{
Kitagawa–Blinder–Oaxaca decomposition
Author: Mediacom
This function allows you to run a summary of Kitagawa–Blinder–Oaxaca decomposition.
summary.oaxaca()
}
